+++ src/main/java/com/cta4j/bus/model/BusPredictionType.java
@@ -1,0 +1,16 @@
+package com.cta4j.bus.model;
+
+/**
+ * A type of bus prediction, either an arrival or a departure.
+ */
+public enum BusPredictionType {
+    /**
+     * Indicates an arrival prediction.
+     */
+    ARRIVAL,
+
+    /**
+     * Indicates a departure prediction.
+     */
+    DEPARTURE
+}