+++ src/main/java/com/cta4j/train/model/Route.java
@@ -1,0 +1,51 @@
+package com.cta4j.train.model;
+
+/**
+ * A train route.
+ */
+public enum Route {
+    /**
+     * The Red Line route.
+     */
+    RED,
+
+    /**
+     * The Blue Line route.
+     */
+    BLUE,
+
+    /**
+     * The Brown Line route.
+     */
+    BROWN,
+
+    /**
+     * The Green Line route.
+     */
+    GREEN,
+
+    /**
+     * The Orange Line route.
+     */
+    ORANGE,
+
+    /**
+     * The Purple Line route.
+     */
+    PURPLE,
+
+    /**
+     * The Pink Line route.
+     */
+    PINK,
+
+    /**
+     * The Yellow Line route.
+     */
+    YELLOW,
+
+    /**
+     * Not available / not applicable.
+     */
+    N_A
+}